refactored tests; fixed merge cnoflicts
